Move client initialization code to separate nice file
